moved automatic target distribution generation to density_estimation.py
